newrequirements.txt
Django>=2.2.9
django-cors-headers>=2.4.0
django-filter>=2.1.0
djangorestframework>=3.9.1
djangorestframework-jwt==1.11.0
Markdown==3.0.1
PyJWT==1.7.1
pytz==2018.9
psutil==5.5.0
djangorestframework-csv==2.1.0